Permanently move JQ out of sandbox
