+++ src/sandbox/messenger/registration.ts
@@ -18,15 +18,10 @@
 /** @file It doesn't actually use the Messenger but this file tries to replicate the pattern */
 
 import { addPostMessageListener } from "@/utils/postMessage";
-import {
-  applyJq,
-  renderHandlebarsTemplate,
-  renderNunjucksTemplate,
-} from "./executor";
+import { renderHandlebarsTemplate, renderNunjucksTemplate } from "./executor";
 
 export default function registerMessenger(): void {
   addPostMessageListener("SANDBOX_PING", async (payload) => "pong");
   addPostMessageListener("RENDER_NUNJUCKS", renderNunjucksTemplate);
   addPostMessageListener("RENDER_HANDLEBARS", renderHandlebarsTemplate);
-  addPostMessageListener("APPLY_JQ", applyJq);
 }
